competiton auton updates

+++ include/Roller.hpp
@@ -18,5 +18,5 @@ public:
 	Roller(int8_t const port);
 	void switch_color() const;
 	void fine_adjust(RollerCode code);
-	void spin_wheel(int scaler) ;	
+	void spin_wheel(double scaler) ;	
 };
